requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
APScheduler==3.11.0
asttokens==3.0.0
asyncer==0.0.8
attrs==24.3.0
backoff==2.2.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.0
colorlog==6.9.0
cryptography==42.0.8
datasets==3.2.0
decorator==5.1.1
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
dspy==2.5.43
email_validator==2.2.0
executing==2.1.0
fastapi==0.111.1
fastapi-cli==0.0.7
fastapi-sso==0.10.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
importlib_metadata==8.5.0
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json_repair==0.33.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.53.7
magicattr==0.1.6
Mako==1.3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.19.1
numpy==2.2.1
oauthlib==3.2.2
openai==1.58.1
optuna==4.1.0
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
PyNaCl==1.5.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.12.0
rpds-py==0.22.3
rq==2.1.0
rsa==4.9
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.37.2
streamlit==1.41.1
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.22.0
uvloop==0.21.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
